[import] Rename LoaderError to ImportError
