+++ src/import/ImportError.cpp
@@ -42,40 +42,16 @@
 //  statement from your version.
 //
 
-#ifndef _LIB_CASMFE_LOADER_ERROR_H_
-#define _LIB_CASMFE_LOADER_ERROR_H_
+#include "ImportError.h"
 
-#include <exception>
-#include <string>
+using namespace libcasm_fe;
 
-namespace libcasm_fe
+ImportError::ImportError( const std::string& msg )
+: m_msg( msg )
 {
-    class LoaderError : public std::exception
-    {
-      public:
-        explicit LoaderError( const std::string& msg );
-
-        const char* what( void ) const noexcept override;
-
-      private:
-        const std::string m_msg;
-    };
-
-    class NoSuchSpecificationError : public LoaderError
-    {
-      public:
-        using LoaderError::LoaderError;
-    };
 }
 
-#endif  // _LIB_CASMFE_LOADER_ERROR_H_
-
-//
-//  Local variables:
-//  mode: c++
-//  indent-tabs-mode: nil
-//  c-basic-offset: 4
-//  tab-width: 4
-//  End:
-//  vim:noexpandtab:sw=4:ts=4:
-//
+const char* ImportError::what( void ) const noexcept
+{
+    return m_msg.c_str();
+}
